Start feature-focused decompiler fixtures
Document the placement convention and move the spilled receiver/coalesce compiler samples out of the shared CFG fixture without changing their source shape.
